multimedia_system/public_html/yolo/runYolo/requirements.txt
numpy==1.26.4
flask
gunicorn
google-cloud-storage
opencv-python-headless
cloud-sql-python-connector[pg8000]
ultralytics==8.1.19
torch==2.2.2
